Add files via upload
Here's my second commit again, but this time it is in an ASM file. I will be pushing the rest of my commits through this file. I replaced .data with .text because this will help me with my code as of right now. I am having trouble figuring out how to print an integer right now so I commented out the code call for integers, and I'm moving on to my first name for the time being. I started by printing the first character 'K'.